add iTunes Search API for search data

+++ lib/common/models/discover_model/search_result_model.dart
@@ -85,8 +85,8 @@ class Result {
   String? artworkUrl30;
   String? artworkUrl60;
   String? artworkUrl100;
-  int? collectionPrice;
-  int? trackPrice;
+  double? collectionPrice;
+  double? trackPrice;
   int? trackRentalPrice;
   int? collectionHdPrice;
   int? trackHdPrice;
